tests/wpt/web-platform-tests/webstorage/storage_supported_property_names.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Storage Test: Supported property names</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
    ["localStorage", "sessionStorage"].forEach(function(name) {
        test(function() {
            var storage = window[name];
            storage.clear();

            storage["name"] = "user1";
            assert_array_equals(Object.getOwnPropertyNames(storage), ['name']);
        }, "Object.getOwnPropertyNames on " + name + " Storage");

        test(function() {
            var storage = window[name];
            storage.clear();
            assert_array_equals(Object.getOwnPropertyNames(storage), []);
        }, "Object.getOwnPropertyNames on " + name + " storage with empty collection");
    });
</script>
